+++ src/main/kotlin/app/revanced/patches/ad/VideoAdsPatch.kt
@@ -1,14 +1,14 @@
 package app.revanced.patches.ad
 
 import app.revanced.patcher.PatcherData
-import app.revanced.patcher.extensions.AccessFlagExtensions.Companion.or
+import app.revanced.patcher.extensions.or
 import app.revanced.patcher.extensions.addInstructions
 import app.revanced.patcher.patch.*
 import app.revanced.patcher.signature.MethodMetadata
 import app.revanced.patcher.signature.MethodSignature
 import app.revanced.patcher.signature.MethodSignatureMetadata
 import app.revanced.patcher.signature.PatternScanMethod
-import app.revanced.patcher.smali.asInstructions
+import app.revanced.patcher.smali.toInstructions
 import org.jf.dexlib2.AccessFlags
 import org.jf.dexlib2.Opcode
 
@@ -94,7 +94,7 @@ class VideoAdsPatch : Patch(
             """
                 invoke-static { }, Lfi/vanced/libraries/youtube/whitelisting/Whitelist;->shouldShowAds()Z
                 move-result v1
-            """.trimIndent().asInstructions()
+            """.trimIndent().toInstructions()
         )
 
         return PatchResultSuccess()
